Govardovskii et al. (2000) visual-pigment template.

Reconstructs a photoreceptor's normalised absorbance spectrum from the peak wavelength alone, using the parametric template of Govardovskii et al. (2000) for A1 and A2 chromophores. This is the standard workhorse of comparative visual physiology.

Usage

govardovskii_template(lambda, lambda_max, chromophore = c("A1", "A2"))

Arguments

lambda

Wavelength grid in nm (numeric vector).

lambda_max

Peak absorbance wavelength in nm.

chromophore

Chromophore type: "A1" (11-cis retinal, default) or "A2" (11-cis 3,4-didehydroretinal). A2 shifts the peak toward longer wavelengths and broadens the spectrum.

Value

A data frame with columns lambda and S (normalised absorbance, 0–1).

Details

The template models absorbance as the sum of an \(\alpha\)-band (main peak) and a \(\beta\)-band (short-wavelength shoulder). The \(\alpha\)-band follows the four-parameter expression: $$S_\alpha(\lambda) = \left[ e^{A(a - x)} + e^{B(0.922 - x)} + e^{C(1.104 - x)} + D \right]^{-1}, \quad x = \frac{\lambda_{\max}}{\lambda}$$ where \(A\), \(B\), \(C\), \(D\), and \(a\) are chromophore-specific constants (A1: \(A=69.7\), \(B=28\), \(C=-14.9\), \(D=0.674\), \(a\) wavelength-dependent; A2: \(A=62.7\), \(B=26.5\), \(C=-15.6\), \(D=0.530\)). The \(\beta\)-band is a Gaussian: $$S_\beta(\lambda) = 0.26\,\exp\!\left[ -\!\left(\frac{\lambda - \lambda_\beta}{b}\right)^2 \right]$$ with \(\lambda_\beta = 189 + 0.315\,\lambda_{\max}\) and \(b = -40.5 + 0.195\,\lambda_{\max}\). The full template is \(S(\lambda) = \max(0,\, S_\alpha + S_\beta)\), normalised to a peak of 1.

References

Govardovskii VI, Fyhrquist N, Reuter T, Kuzmin DG, Donner K (2000) In search of the visual pigment template. Visual Neuroscience 17:509–528.

Examples

lam <- seq(300, 750, by = 1)
govardovskii_template(lam, lambda_max = 560)
#>     lambda            S
#> 1      300 1.049643e-01
#> 2      301 1.078918e-01
#> 3      302 1.108541e-01
#> 4      303 1.138495e-01
#> 5      304 1.168764e-01
#> 6      305 1.199330e-01
#> 7      306 1.230175e-01
#> 8      307 1.261279e-01
#> 9      308 1.292623e-01
#> 10     309 1.324185e-01
#> 11     310 1.355944e-01
#> 12     311 1.387878e-01
#> 13     312 1.419963e-01
#> 14     313 1.452176e-01
#> 15     314 1.484492e-01
#> 16     315 1.516886e-01
#> 17     316 1.549331e-01
#> 18     317 1.581802e-01
#> 19     318 1.614272e-01
#> 20     319 1.646712e-01
#> 21     320 1.679094e-01
#> 22     321 1.711391e-01
#> 23     322 1.743573e-01
#> 24     323 1.775610e-01
#> 25     324 1.807473e-01
#> 26     325 1.839131e-01
#> 27     326 1.870555e-01
#> 28     327 1.901714e-01
#> 29     328 1.932578e-01
#> 30     329 1.963115e-01
#> 31     330 1.993295e-01
#> 32     331 2.023086e-01
#> 33     332 2.052460e-01
#> 34     333 2.081383e-01
#> 35     334 2.109828e-01
#> 36     335 2.137762e-01
#> 37     336 2.165157e-01
#> 38     337 2.191982e-01
#> 39     338 2.218208e-01
#> 40     339 2.243808e-01
#> 41     340 2.268752e-01
#> 42     341 2.293013e-01
#> 43     342 2.316565e-01
#> 44     343 2.339380e-01
#> 45     344 2.361433e-01
#> 46     345 2.382700e-01
#> 47     346 2.403157e-01
#> 48     347 2.422780e-01
#> 49     348 2.441548e-01
#> 50     349 2.459439e-01
#> 51     350 2.476434e-01
#> 52     351 2.492513e-01
#> 53     352 2.507660e-01
#> 54     353 2.521857e-01
#> 55     354 2.535088e-01
#> 56     355 2.547341e-01
#> 57     356 2.558602e-01
#> 58     357 2.568859e-01
#> 59     358 2.578103e-01
#> 60     359 2.586325e-01
#> 61     360 2.593518e-01
#> 62     361 2.599675e-01
#> 63     362 2.604793e-01
#> 64     363 2.608869e-01
#> 65     364 2.611901e-01
#> 66     365 2.613889e-01
#> 67     366 2.614835e-01
#> 68     367 2.614742e-01
#> 69     368 2.613615e-01
#> 70     369 2.611461e-01
#> 71     370 2.608286e-01
#> 72     371 2.604101e-01
#> 73     372 2.598916e-01
#> 74     373 2.592744e-01
#> 75     374 2.585598e-01
#> 76     375 2.577495e-01
#> 77     376 2.568450e-01
#> 78     377 2.558482e-01
#> 79     378 2.547611e-01
#> 80     379 2.535858e-01
#> 81     380 2.523246e-01
#> 82     381 2.509798e-01
#> 83     382 2.495540e-01
#> 84     383 2.480499e-01
#> 85     384 2.464701e-01
#> 86     385 2.448176e-01
#> 87     386 2.430955e-01
#> 88     387 2.413068e-01
#> 89     388 2.394547e-01
#> 90     389 2.375427e-01
#> 91     390 2.355742e-01
#> 92     391 2.335527e-01
#> 93     392 2.314819e-01
#> 94     393 2.293654e-01
#> 95     394 2.272071e-01
#> 96     395 2.250109e-01
#> 97     396 2.227808e-01
#> 98     397 2.205207e-01
#> 99     398 2.182347e-01
#> 100    399 2.159271e-01
#> 101    400 2.136019e-01
#> 102    401 2.112635e-01
#> 103    402 2.089162e-01
#> 104    403 2.065642e-01
#> 105    404 2.042121e-01
#> 106    405 2.018641e-01
#> 107    406 1.995247e-01
#> 108    407 1.971984e-01
#> 109    408 1.948896e-01
#> 110    409 1.926028e-01
#> 111    410 1.903426e-01
#> 112    411 1.881133e-01
#> 113    412 1.859196e-01
#> 114    413 1.837659e-01
#> 115    414 1.816567e-01
#> 116    415 1.795965e-01
#> 117    416 1.775897e-01
#> 118    417 1.756409e-01
#> 119    418 1.737544e-01
#> 120    419 1.719347e-01
#> 121    420 1.701861e-01
#> 122    421 1.685130e-01
#> 123    422 1.669198e-01
#> 124    423 1.654106e-01
#> 125    424 1.639899e-01
#> 126    425 1.626617e-01
#> 127    426 1.614303e-01
#> 128    427 1.602998e-01
#> 129    428 1.592742e-01
#> 130    429 1.583577e-01
#> 131    430 1.575541e-01
#> 132    431 1.568675e-01
#> 133    432 1.563016e-01
#> 134    433 1.558604e-01
#> 135    434 1.555475e-01
#> 136    435 1.553667e-01
#> 137    436 1.553216e-01
#> 138    437 1.554157e-01
#> 139    438 1.556526e-01
#> 140    439 1.560357e-01
#> 141    440 1.565684e-01
#> 142    441 1.572538e-01
#> 143    442 1.580953e-01
#> 144    443 1.590959e-01
#> 145    444 1.602587e-01
#> 146    445 1.615866e-01
#> 147    446 1.630825e-01
#> 148    447 1.647491e-01
#> 149    448 1.665891e-01
#> 150    449 1.686051e-01
#> 151    450 1.707996e-01
#> 152    451 1.731748e-01
#> 153    452 1.757332e-01
#> 154    453 1.784767e-01
#> 155    454 1.814074e-01
#> 156    455 1.845273e-01
#> 157    456 1.878380e-01
#> 158    457 1.913412e-01
#> 159    458 1.950384e-01
#> 160    459 1.989310e-01
#> 161    460 2.030201e-01
#> 162    461 2.073069e-01
#> 163    462 2.117921e-01
#> 164    463 2.164767e-01
#> 165    464 2.213610e-01
#> 166    465 2.264456e-01
#> 167    466 2.317306e-01
#> 168    467 2.372161e-01
#> 169    468 2.429020e-01
#> 170    469 2.487879e-01
#> 171    470 2.548733e-01
#> 172    471 2.611574e-01
#> 173    472 2.676393e-01
#> 174    473 2.743179e-01
#> 175    474 2.811919e-01
#> 176    475 2.882595e-01
#> 177    476 2.955191e-01
#> 178    477 3.029686e-01
#> 179    478 3.106059e-01
#> 180    479 3.184283e-01
#> 181    480 3.264332e-01
#> 182    481 3.346178e-01
#> 183    482 3.429788e-01
#> 184    483 3.515128e-01
#> 185    484 3.602162e-01
#> 186    485 3.690852e-01
#> 187    486 3.781157e-01
#> 188    487 3.873034e-01
#> 189    488 3.966437e-01
#> 190    489 4.061319e-01
#> 191    490 4.157630e-01
#> 192    491 4.255317e-01
#> 193    492 4.354328e-01
#> 194    493 4.454605e-01
#> 195    494 4.556091e-01
#> 196    495 4.658725e-01
#> 197    496 4.762445e-01
#> 198    497 4.867187e-01
#> 199    498 4.972885e-01
#> 200    499 5.079473e-01
#> 201    500 5.186880e-01
#> 202    501 5.295035e-01
#> 203    502 5.403868e-01
#> 204    503 5.513304e-01
#> 205    504 5.623267e-01
#> 206    505 5.733682e-01
#> 207    506 5.844472e-01
#> 208    507 5.955557e-01
#> 209    508 6.066857e-01
#> 210    509 6.178293e-01
#> 211    510 6.289783e-01
#> 212    511 6.401243e-01
#> 213    512 6.512592e-01
#> 214    513 6.623745e-01
#> 215    514 6.734617e-01
#> 216    515 6.845124e-01
#> 217    516 6.955180e-01
#> 218    517 7.064698e-01
#> 219    518 7.173592e-01
#> 220    519 7.281775e-01
#> 221    520 7.389160e-01
#> 222    521 7.495659e-01
#> 223    522 7.601184e-01
#> 224    523 7.705646e-01
#> 225    524 7.808957e-01
#> 226    525 7.911029e-01
#> 227    526 8.011771e-01
#> 228    527 8.111094e-01
#> 229    528 8.208909e-01
#> 230    529 8.305127e-01
#> 231    530 8.399656e-01
#> 232    531 8.492407e-01
#> 233    532 8.583289e-01
#> 234    533 8.672211e-01
#> 235    534 8.759082e-01
#> 236    535 8.843811e-01
#> 237    536 8.926307e-01
#> 238    537 9.006478e-01
#> 239    538 9.084232e-01
#> 240    539 9.159478e-01
#> 241    540 9.232123e-01
#> 242    541 9.302075e-01
#> 243    542 9.369242e-01
#> 244    543 9.433532e-01
#> 245    544 9.494854e-01
#> 246    545 9.553116e-01
#> 247    546 9.608226e-01
#> 248    547 9.660094e-01
#> 249    548 9.708630e-01
#> 250    549 9.753743e-01
#> 251    550 9.795347e-01
#> 252    551 9.833353e-01
#> 253    552 9.867676e-01
#> 254    553 9.898231e-01
#> 255    554 9.924936e-01
#> 256    555 9.947711e-01
#> 257    556 9.966477e-01
#> 258    557 9.981159e-01
#> 259    558 9.991686e-01
#> 260    559 9.997988e-01
#> 261    560 1.000000e+00
#> 262    561 9.997662e-01
#> 263    562 9.990917e-01
#> 264    563 9.979714e-01
#> 265    564 9.964007e-01
#> 266    565 9.943755e-01
#> 267    566 9.918923e-01
#> 268    567 9.889485e-01
#> 269    568 9.855420e-01
#> 270    569 9.816713e-01
#> 271    570 9.773359e-01
#> 272    571 9.725360e-01
#> 273    572 9.672727e-01
#> 274    573 9.615478e-01
#> 275    574 9.553641e-01
#> 276    575 9.487255e-01
#> 277    576 9.416364e-01
#> 278    577 9.341026e-01
#> 279    578 9.261304e-01
#> 280    579 9.177275e-01
#> 281    580 9.089022e-01
#> 282    581 8.996640e-01
#> 283    582 8.900231e-01
#> 284    583 8.799908e-01
#> 285    584 8.695793e-01
#> 286    585 8.588015e-01
#> 287    586 8.476712e-01
#> 288    587 8.362030e-01
#> 289    588 8.244122e-01
#> 290    589 8.123148e-01
#> 291    590 7.999275e-01
#> 292    591 7.872673e-01
#> 293    592 7.743521e-01
#> 294    593 7.611998e-01
#> 295    594 7.478289e-01
#> 296    595 7.342583e-01
#> 297    596 7.205067e-01
#> 298    597 7.065935e-01
#> 299    598 6.925377e-01
#> 300    599 6.783585e-01
#> 301    600 6.640750e-01
#> 302    601 6.497062e-01
#> 303    602 6.352708e-01
#> 304    603 6.207873e-01
#> 305    604 6.062740e-01
#> 306    605 5.917485e-01
#> 307    606 5.772282e-01
#> 308    607 5.627301e-01
#> 309    608 5.482704e-01
#> 310    609 5.338650e-01
#> 311    610 5.195290e-01
#> 312    611 5.052772e-01
#> 313    612 4.911235e-01
#> 314    613 4.770812e-01
#> 315    614 4.631630e-01
#> 316    615 4.493810e-01
#> 317    616 4.357464e-01
#> 318    617 4.222701e-01
#> 319    618 4.089619e-01
#> 320    619 3.958314e-01
#> 321    620 3.828871e-01
#> 322    621 3.701374e-01
#> 323    622 3.575896e-01
#> 324    623 3.452507e-01
#> 325    624 3.331270e-01
#> 326    625 3.212244e-01
#> 327    626 3.095480e-01
#> 328    627 2.981027e-01
#> 329    628 2.868926e-01
#> 330    629 2.759217e-01
#> 331    630 2.651932e-01
#> 332    631 2.547099e-01
#> 333    632 2.444745e-01
#> 334    633 2.344890e-01
#> 335    634 2.247551e-01
#> 336    635 2.152742e-01
#> 337    636 2.060471e-01
#> 338    637 1.970746e-01
#> 339    638 1.883569e-01
#> 340    639 1.798939e-01
#> 341    640 1.716854e-01
#> 342    641 1.637306e-01
#> 343    642 1.560286e-01
#> 344    643 1.485779e-01
#> 345    644 1.413770e-01
#> 346    645 1.344241e-01
#> 347    646 1.277167e-01
#> 348    647 1.212526e-01
#> 349    648 1.150287e-01
#> 350    649 1.090421e-01
#> 351    650 1.032894e-01
#> 352    651 9.776699e-02
#> 353    652 9.247088e-02
#> 354    653 8.739697e-02
#> 355    654 8.254085e-02
#> 356    655 7.789791e-02
#> 357    656 7.346330e-02
#> 358    657 6.923199e-02
#> 359    658 6.519876e-02
#> 360    659 6.135824e-02
#> 361    660 5.770489e-02
#> 362    661 5.423308e-02
#> 363    662 5.093704e-02
#> 364    663 4.781096e-02
#> 365    664 4.484893e-02
#> 366    665 4.204503e-02
#> 367    666 3.939330e-02
#> 368    667 3.688780e-02
#> 369    668 3.452261e-02
#> 370    669 3.229184e-02
#> 371    670 3.018967e-02
#> 372    671 2.821035e-02
#> 373    672 2.634824e-02
#> 374    673 2.459779e-02
#> 375    674 2.295357e-02
#> 376    675 2.141030e-02
#> 377    676 1.996282e-02
#> 378    677 1.860613e-02
#> 379    678 1.733539e-02
#> 380    679 1.614593e-02
#> 381    680 1.503324e-02
#> 382    681 1.399298e-02
#> 383    682 1.302098e-02
#> 384    683 1.211327e-02
#> 385    684 1.126602e-02
#> 386    685 1.047560e-02
#> 387    686 9.738545e-03
#> 388    687 9.051547e-03
#> 389    688 8.411478e-03
#> 390    689 7.815367e-03
#> 391    690 7.260401e-03
#> 392    691 6.743920e-03
#> 393    692 6.263413e-03
#> 394    693 5.816511e-03
#> 395    694 5.400981e-03
#> 396    695 5.014723e-03
#> 397    696 4.655763e-03
#> 398    697 4.322247e-03
#> 399    698 4.012436e-03
#> 400    699 3.724700e-03
#> 401    700 3.457512e-03
#> 402    701 3.209443e-03
#> 403    702 2.979159e-03
#> 404    703 2.765409e-03
#> 405    704 2.567031e-03
#> 406    705 2.382935e-03
#> 407    706 2.212108e-03
#> 408    707 2.053605e-03
#> 409    708 1.906546e-03
#> 410    709 1.770110e-03
#> 411    710 1.643534e-03
#> 412    711 1.526110e-03
#> 413    712 1.417176e-03
#> 414    713 1.316120e-03
#> 415    714 1.222371e-03
#> 416    715 1.135400e-03
#> 417    716 1.054715e-03
#> 418    717 9.798592e-04
#> 419    718 9.104091e-04
#> 420    719 8.459713e-04
#> 421    720 7.861810e-04
#> 422    721 7.306999e-04
#> 423    722 6.792138e-04
#> 424    723 6.314319e-04
#> 425    724 5.870842e-04
#> 426    725 5.459206e-04
#> 427    726 5.077091e-04
#> 428    727 4.722347e-04
#> 429    728 4.392983e-04
#> 430    729 4.087151e-04
#> 431    730 3.803142e-04
#> 432    731 3.539368e-04
#> 433    732 3.294362e-04
#> 434    733 3.066761e-04
#> 435    734 2.855303e-04
#> 436    735 2.658821e-04
#> 437    736 2.476231e-04
#> 438    737 2.306528e-04
#> 439    738 2.148784e-04
#> 440    739 2.002135e-04
#> 441    740 1.865783e-04
#> 442    741 1.738988e-04
#> 443    742 1.621064e-04
#> 444    743 1.511374e-04
#> 445    744 1.409330e-04
#> 446    745 1.314384e-04
#> 447    746 1.226031e-04
#> 448    747 1.143799e-04
#> 449    748 1.067255e-04
#> 450    749 9.959933e-05
#> 451    750 9.296400e-05